# src/main/java/es/iesjandula/adrian_luna_video_club/models/Movie.java
package es.iesjandula.adrian_luna_video_club.models;

import java.util.List;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.OneToMany;
import jakarta.persistence.Table;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Entity
@Table(name = "movies")
public class Movie
{
	@Id
	@GeneratedValue(strategy = GenerationType.IDENTITY)
	@Column
	private Long movieId;

	@Column(nullable = false)
	private String title;
	
	@Column(nullable = false)
	private String duration;
	
	@Column(nullable = false)
	private Long stock;
	
	
	
	@OneToMany(mappedBy = "movie")
    private List<Booking> bookings;
}