# src/main/resources/Convertir Swagger en proyecto Spring Boot - Videoclub online.yaml
swagger: '2.0'
host: localhost:8080
basePath: /videoclub
info:
  title: videoclub
  version: 1.0.0
  description: Esta especificación muestra cómo gestionar un videoclub
  contact:
    name: Fulanito Menganito
    email: fulanito.menganito@empresa.com
    url: https://www.videoclub.es
schemes:
 - http
 
          ######################################################
          ############### ENTRYPOINTS y ENDPOINTS ##############
          ######################################################
paths:
...
  /booking:

    get:
      operationId: getMovies
      description: Obtener la lista de películas almacenadas en sesión (no alquiladas aún)
      produces:
        - application/json
      responses:
        200:
          description: Devuelve la lista de películas de películas almacenadas en sesión (no alquiladas aún)
          schema:
            type: array
            items:
              $ref: '#/definitions/Movie'
          examples:
            application/json: [ 
              { "movieId": 1, "title": "Encanto", "duration" : "01:42" },
              { "movieId": 2, "title": "Frozen: El reino del hielo", "duration" : "01:41" },
              { "movieId": 3, "title": "Hércules", "duration" : "01:33" },
              { "movieId": 4, "title": "Vaiana", "duration" : "01:47" }
            ]

    post:
      operationId: searchMovies
      description: Obtener la lista de películas almacenadas en sesión que \
        cumplen alguno de los parámetros que se envían (no alquiladas aún)
      parameters:
        - name: searchByFields
          in: body
          description: objeto que lleva rellenos aquellos campos con los que buscar
          required: true
          schema:
            $ref: '#/definitions/Movie'
      produces:
        - application/json
      responses:
        200:
          description: Devuelve una lista de películas almacenadas en sesión que \
            cumplen los criterios de búsqueda (no alquiladas aún)
          schema:
            type: array
            items:
              $ref: '#/definitions/Movie'
          examples:
            application/json: [ 
              { "movieId": 1, "title": "Encanto", "duration" : "01:42" }
            ]
  /booking/{movieId}:
  
    parameters:
      - name: movieId
        in: path
        description: Identificador de la película
        required: true
        type: integer
        format: int64
      - name: userId
        in: query
        description: Identificador del usuario
        required: true
        type: integer
        format: int64
            
    post:
      operationId: bookingMovie
      description: Reservar película a un usuario
      produces:
        - application/json
      responses:
        200:
          description: Película reservada correctamente
        400:
          description: Película no encontrada
          schema:
            $ref: '#/definitions/MovieError'
          examples:
            application/json: { "movieErrorId": 1, "message": "Movie was not found" }
        404:
          description: El usuario no existe
          schema:
            $ref: '#/definitions/MovieError'   
          examples:
            application/json: { "movieErrorId": 1, "message": "User was not found" }

    delete:
      operationId: cancelBookingMovie
      description: Cancelar película de un usuario
      produces:
        - application/json
      responses:
        200:
          description: Cancelación de reservada realizada
        400:
          description: Película no encontrada
          schema:
            $ref: '#/definitions/MovieError'
          examples:
            application/json: { "movieErrorId": 1, "message": "Movie was not found" }
        404:
          description: El usuario no existe
          schema:
            $ref: '#/definitions/MovieError'   
          examples:
            application/json: { "movieErrorId": 1, "message": "User was not found" }


            ##################################################
            ######## TAREAS Thymeleaf (EntryPoints) ##########
            ##################################################
...
definitions:
  Movie:
    type: object
    properties:
      movieId:
        type: integer
        format: int64
        description: Identificador de la película
      title:
        type: string
        description: Título de la película
      duration:
        type: string
        description: Duración de la película en formato HH:MM
...
  MovieError:
    type: object
    properties:
      movieErrorId:
        type: integer
        format: int64
        description: Identificador de error
      message:
        type: string
        description: Mensaje de error
...